Add Edit

+++ UsedCar/src/java/controller/KKPUsedControl.java
@@ -42,7 +42,7 @@ public class KKPUsedControl extends HttpServlet {
         String guarantee = request.getParameter("guarantee");
         float DBR = Float.parseFloat(request.getParameter("DBR"));
         float IIR = Float.parseFloat(request.getParameter("IIR"));
-        long exceptIncome =0;
+        long exceptIncome = 0;
         if (Integer.parseInt(guarantee) == 0) {// ถ้ายกเว้นผู้ค้ำประกันได้
             //ยกเว้นผู้ค้ำประกันได้ ถ้ารายได้มากกว่า 30,000 บาท หรือ ปรับ LTV<=90% เทอมไม่เกิน 72	            
             DecimalFormat df = new DecimalFormat("##,###,###,###");
@@ -56,10 +56,10 @@ public class KKPUsedControl extends HttpServlet {
             guarantee = "ต้องมีผู้ค้ำประกัน";
         }
         String NCB_Name = NCB.getNCBNameFromType(NCB_Type);
-        boolean result = pdpg_used.addPDPG_USED(gradeTentId, NCB_Type, NCB_Name, occupation, carAgeId, maxLTV, maxTerm, DBR, IIR, exceptIncome, guarantee);
+        boolean result = pdpg_used.addPDPG_Used(gradeTentId, NCB_Type, NCB_Name, occupation, carAgeId, maxLTV, maxTerm, DBR, IIR, exceptIncome, guarantee);
         String notiHead = "Action Result";
         String notiMessage = "";
-        if(result){
+        if (result) {
             notiMessage = "การเพิ่มสำเร็จ";
         } else {
             notiMessage = "การเพิ่มไม่สำเร็จ";
@@ -67,7 +67,9 @@ public class KKPUsedControl extends HttpServlet {
         request.setAttribute("notiHead", notiHead);
         request.setAttribute("notiMessage", notiMessage);
         request.setAttribute("showNoti", "true");
+        request.setAttribute("refreshPage", request.getContextPath() + "/Update");
         getServletContext().getRequestDispatcher("/Update").forward(request, response);
+//        response.sendRedirect(request.getContextPath() + "/Update");
     }
 
     // <editor-fold defaultstate="collapsed" desc="HttpServlet methods. Click on the + sign on the left to edit the code.">
@@ -82,7 +84,12 @@ public class KKPUsedControl extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        processRequest(request, response);
+        if (request.getParameter("gradeTent") != null) { // if doGet and have data
+            processRequest(request, response);
+        } else { // if doGet and don't have data 
+            response.sendRedirect(request.getContextPath() + "/Update");
+        }
+
     }
 
     /**
@@ -96,13 +103,46 @@ public class KKPUsedControl extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        int deleteId = Integer.parseInt(request.getParameter("id"));        
-        if(pdpg_used.deletePDPG_UsedData(deleteId)){
-            //success
+        request.setCharacterEncoding("UTF-8");
+        String target = request.getParameter("target");        
+        if (target != null) {
+            if (target.equalsIgnoreCase("delete")) {
+                int deleteId = Integer.parseInt(request.getParameter("id"));
+                if (pdpg_used.deletePDPG_UsedData(deleteId)) {
+                    //success
+                } else {
+                    //not success
+                }
+            }
         } else {
-            //not success
+            int targetId = Integer.parseInt(request.getParameter("dataId"));
+            int gradeTentId = Integer.parseInt(request.getParameter("gradeTent"));
+            int NCB_Type = Integer.parseInt(request.getParameter("NCBType"));
+            String occupation = request.getParameter("occupation");
+            int carAgeId = Integer.parseInt(request.getParameter("carAge"));
+            float maxLTV = Float.parseFloat(request.getParameter("maxLTV"));
+            float maxTerm = Float.parseFloat(request.getParameter("maxTerm"));
+            String guarantee = request.getParameter("guarantee");
+            float DBR = Float.parseFloat(request.getParameter("DBR"));
+            float IIR = Float.parseFloat(request.getParameter("IIR"));
+            long exceptIncome = Long.parseLong(request.getParameter("exceptIncome"));
+            String NCB_Name = NCB.getNCBNameFromType(NCB_Type);
+            boolean result = pdpg_used.updatePDPG_Used(targetId, gradeTentId, NCB_Type, NCB_Name, occupation, carAgeId, maxLTV, maxTerm, DBR, IIR, exceptIncome, guarantee);
+            String notiHead = "Action Result";
+            String notiMessage = "";
+            if (result) {
+                notiMessage = "การแก้ไขข้อมูลสำเร็จ";
+            } else {
+                notiMessage = "การแก้ไขข้อมูลไม่สำเร็จ";
+            }
+            request.setAttribute("notiHead", notiHead);
+            request.setAttribute("notiMessage", notiMessage);
+            request.setAttribute("showNoti", "true");
+            request.setAttribute("refreshPage", request.getContextPath() + "/Update");
+            getServletContext().getRequestDispatcher("/Update").forward(request, response);
+//            response.sendRedirect(request.getContextPath() + "/Update");
         }
-        
+
     }
 
     /**
